Converted all inputs to floats to prevent overflow/underflow

+++ main.cpp
@@ -7,23 +7,23 @@ void print( std::string txt ) {
 	std::cout << txt << std::endl;
 }
 
-double input( std::string prefix ) {
-	double returnVal;
+float input( std::string prefix ) {
+	float returnVal;
 	std::cout << prefix;
 	std::cin >> returnVal;
 
 	return returnVal;
 }
 
-std::array<double, 3> pqFormel( double a, double b, double c ) {
-	double p = b/a;
-	double q = c/a;
+std::array<float, 3> pqFormel( float a, float b, float c ) {
+	float p = b/a;
+	float q = c/a;
 
-	double base = -(p/2);
-	double rotDiff = abs( pow(base, 2) - q );
-	double rot = sqrt( rotDiff );
+	float base = -(p/2);
+	float rotDiff = abs( pow(base, 2) - q );
+	float rot = sqrt( rotDiff );
 
-	double isComplex = 0.0;
+	float isComplex = 0.0;
 	if( pow(base, 2) - q < 0 ) {
 		isComplex = 1.0;
 	}
@@ -31,7 +31,7 @@ std::array<double, 3> pqFormel( double a, double b, double c ) {
 	return { base, rot, isComplex };
 }
 
-std::string stringResult( std::array<double, 3> res ) {
+std::string stringResult( std::array<float, 3> res ) {
 	std::string endStr = "";
 	if( res[2] == 1.0 ) {
 		endStr = "i";
@@ -42,20 +42,20 @@ std::string stringResult( std::array<double, 3> res ) {
 
 class Equation {
 	private:
-		double a = 1;
-		double b = 0;
-		double c = 0;
+		float a = 1;
+		float b = 0;
+		float c = 0;
 	public:
-		std::array<double, 3> solution;
+		std::array<float, 3> solution;
 		std::string solutionStr = "Not solved";
-		Equation( double A, double B, double C ) {
+		Equation( float A, float B, float C ) {
 			a = A;
 			b = B;
 			c = C;
 		}
 
 		void solve() {
-			std::array<double, 3> res = pqFormel( a, b, c );
+			std::array<float, 3> res = pqFormel( a, b, c );
 			solutionStr = stringResult(res);
 		}
 };
@@ -64,14 +64,14 @@ int main() {
 	print( "A(x^2) + Bx + C = 0" );
 	print( "Input your equation" );
 
-	double A = input("A: ");
+	float A = input("A: ");
 	if( A == 0 ) {
 		print("That is no quadratic function!");
 		return 1;
 	}
 
-	double B = input("B (p): ");
-	double C = input("C (q): ");
+	float B = input("B (p): ");
+	float C = input("C (q): ");
 
 	Equation ekv( A, B, C );
 	ekv.solve();
